pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperion"
version = "3.0.0"
description = "Git Repository Profiler & Knowledge Graph Platform with RAG capabilities"
readme = "README.md"
authors = [
    {name = "Matthieu Ryckman", email = "contact@ryckmat.dev"}
]
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
keywords = ["git", "analysis", "rag", "neo4j", "knowledge-graph", "code-intelligence"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: POSIX :: Linux",
]

dependencies = [
    # Core
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",

    # API REST
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",

    # RAG
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "torch>=2.0.0",

    # Integrations
    "neo4j>=5.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-benchmark>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "radon>=5.1.0",
    "pylint>=3.0.0",
    "rouge-score>=0.1.0",
]

monitoring = [
    "psutil>=5.9.0",
]

security = [
    "cryptography>=41.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
]

ml = [
    "mlflow>=2.8.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "joblib>=1.3.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

all = [
    "hyperion[dev,docs,monitoring,security,ml]",
]

[project.urls]
Homepage = "https://github.com/Ryckmat/hyperion"
Documentation = "https://github.com/Ryckmat/hyperion/docs"
Repository = "https://github.com/Ryckmat/hyperion"
"Bug Tracker" = "https://github.com/Ryckmat/hyperion/issues"
Changelog = "https://github.com/Ryckmat/hyperion/blob/main/docs/CHANGELOG.md"

[project.scripts]
hyperion = "hyperion.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
hyperion = ["../templates/**/*", "../config/**/*"]

# ============================================================================
# BLACK - Code Formatting
# ============================================================================
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # defaults
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# ============================================================================
# RUFF - Linting
# ============================================================================
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/**/*.py" = [
    "ARG",  # unused arguments OK in tests
    "S101", # assert usage OK in tests
]

[tool.ruff.lint.isort]
known-first-party = ["hyperion"]

# ============================================================================
# MYPY - Type Checking
# ============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start permissive, tighten later
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "qdrant_client.*",
    "langchain.*",
    "sentence_transformers.*",
    "neo4j.*",
]
ignore_missing_imports = true

# ============================================================================
# PYTEST - Testing
# ============================================================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=hyperion",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow tests",
    "benchmark: Performance benchmark tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
